+++ jgclark.Summaries/src/summaryHelpers.js
@@ -6,13 +6,16 @@
 //-----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
+import moment from 'moment/min/moment-with-locales'
+import { stringListOrArrayToArray } from '@helpers/dataManipulation'
 import {
+  calcOffsetDateStr,
   getDateFromUnhyphenatedDateString,
   getDateStringFromCalendarFilename,
+  RE_YYYYMMDD_DATE,
   unhyphenateString,
   withinDateRange,
 } from '@helpers/dateTime'
-import { calcOffsetDateStr } from '@helpers/NPdateTime'
 import { clo, logDebug, logInfo, logWarn, logError } from '@helpers/dev'
 import {
   CaseInsensitiveMap,
@@ -50,6 +53,8 @@ export type SummariesConfig = {
   progressHeading: string,
   progressHashtags: Array<string>,
   progressMentions: Array<string>,
+  progressMentionsAverage: Array<string>,
+  progressMentionsTotal: Array<string>,
   progressYesNo: Array<string>,
   progressYesNoChars: string,
 }
@@ -63,7 +68,7 @@ export async function getSummariesSettings(): Promise<any> {
   try {
     // Get settings using ConfigV2
     const v2Config: SummariesConfig = await DataStore.loadJSON('../jgclark.Summaries/settings.json')
-    // clo(v2Config, `${pluginID} settings from V2:`)
+    clo(v2Config, `${pluginID} settings from V2:`)
 
     if (v2Config == null || Object.keys(v2Config).length === 0) {
       throw new Error(`Cannot find settings for '${pluginID}' plugin`)
@@ -94,30 +99,39 @@ export class TMOccurrences {
 
   /**
    * Create a new object, initialising the main valuesMap to the required number of values, as 'NaN', so that we can distinguish zero from no occurrences.
-   * (Unless type 'yesno' )
+   * (Unless type 'yesno')
    * @param {string} term 
    * @param {string} type
-   * @param {string} fromDateStr of type YYYYMMDD
-   * @param {string} toDateStr of type YYYYMMDD
+   * @param {string} fromDateStr of type YYYY-MM-DD
+   * @param {string} toDateStr of type YYYY-MM-DD
    */
   constructor(term: string, type: string, fromDateStr: string, toDateStr: string) {
-    this.term = term
-    this.type = type
-    if (toDateStr && fromDateStr) {
-      // $FlowFixMe(incompatible-call)
-      this.numDays = Calendar.unitsBetween(getDateFromUnhyphenatedDateString(toDateStr), getDateFromUnhyphenatedDateString(fromDateStr), 'day') + 1 // inclusive
-      this.valuesMap = new Map < string, number > ()
-      this.total = 0
-      this.count = 0
-      // Initialise all values to NaN, unless type 'yesno'
-      for (let i = 1; i <= this.numDays; i++) {
-        let thisDateStr = unhyphenateString(calcOffsetDateStr(fromDateStr, `${i}d`))
-        // logDebug('TMOccurrences:constructor', `- date ${thisDateStr}`)
-        this.valuesMap.set(thisDateStr, (this.type == 'yesno') ? 0 : NaN)
+    try {
+      this.term = term
+      this.type = type
+      if (toDateStr && fromDateStr) {
+        // Calc number of days to cover
+        // (Moment's diff function returns a truncated number by default, not rounded, so work around that, in case we're getting 6.9 days because of timezone issues)
+        const momFromDate = new moment(fromDateStr, 'YYYY-MM-DD')
+        const momToDate = new moment(toDateStr, 'YYYY-MM-DD')
+        const numDays = Math.round(momToDate.diff(momFromDate, 'days', true)) + 1
+        this.numDays = numDays
+        this.valuesMap = new Map < string, number > ()
+        this.total = 0
+        this.count = 0
+        // Initialise all values to NaN, unless type 'yesno'
+        for (let i = 0; i < numDays; i++) {
+          let thisDateStr = unhyphenateString(calcOffsetDateStr(fromDateStr, `${i}d`))
+          // logDebug('TMOcc:constructor', `- +${i}d -> date ${thisDateStr}`)
+          this.valuesMap.set(thisDateStr, (this.type == 'yesno') ? 0 : NaN)
+        }
+        logDebug('TMOcc:constructor', `Constructed ${term} type ${this.type} for date ${fromDateStr} - ${toDateStr} -> valuesMap for ${this.valuesMap.size} / ${this.numDays} days `)
+      } else {
+        logError('TMOcc:constructor', `Couldn't construct as passed date(s) were empty`)
       }
-      // logDebug('TMOccurrences:constructor', `Constructed ${term} for date ${fromDateStr}-${toDateStr} = ${this.valuesMap.size} days`)
-    } else {
-      logError('TMOccurrences:constructor', `Couldn't construct as passed date(s) were empty`)
+    }
+    catch (error) {
+      logError('TMOcc:constructor', error.message)
     }
   }
 
@@ -127,41 +141,53 @@ export class TMOccurrences {
    * @param {string} dateStr format YYYYMMDD
    */
   addOccurrence(occurrenceStr: string, dateStr: string): void {
-    // logDebug('TMOccurrences:addOccurrence', `starting for ${occurrenceStr} on date = ${dateStr}`)
+    try {
+      if (dateStr == null) {
+        throw new Error(`Passed null date string`)
+      }
+      if (!dateStr.match(RE_YYYYMMDD_DATE)) {
+        throw new Error(`Passed invalid date string '${dateStr}'`)
+      }
+      logDebug('TMOcc:addOccurrence', `starting for ${occurrenceStr} on date = ${dateStr}`)
 
-    // isolate the value
-    let key = occurrenceStr
-    let value = NaN
-    // if this tag that finishes '/number', then break into its two parts, ready to sum the numbers as well
-    // Note: testing includes decimal part of a number, but the API .hashtags drops them
-    if (occurrenceStr.match(/\/-?\d+(\.\d+)?$/)) {
-      const tagParts = occurrenceStr.split('/')
-      key = tagParts[0]
-      value = Number(tagParts[1])
-      // logDebug('TMOccurrences:addOccurrence', `- found tagParts ${key} / ${value.toString()}`)
-    }
-    // if this is a mention that finishes (number), then break into separate parts first
-    else if (occurrenceStr.match(/\(-?\d+(\.\d+)?\)$/)) {
-      const mentionParts = occurrenceStr.split('(')
-      key = mentionParts[0]
-      value = Number.parseFloat(mentionParts[1].slice(0, -1)) // chop off final ')' character
-      // logDebug('TMOccurrences:addOccurrence', `- found tagParts ${key} / ${value.toString()}`)
-    }
+      // isolate the value
+      let key = occurrenceStr
+      let value = NaN
+      // if this tag that finishes '/number', then break into its two parts, ready to sum the numbers as well
+      // Note: testing includes decimal part of a number, but the API .hashtags drops them
+      if (occurrenceStr.match(/\/-?\d+(\.\d+)?$/)) {
+        const tagParts = occurrenceStr.split('/')
+        key = tagParts[0]
+        value = Number(tagParts[1])
+        // logDebug('TMOcc:addOccurrence', `- found tagParts ${key} / ${value.toString()}`)
+      }
+      // if this is a mention that finishes '(number)', then break into separate parts first
+      else if (occurrenceStr.match(/\(-?\d+(\.\d+)?\)$/)) {
+        const mentionParts = occurrenceStr.split('(')
+        key = mentionParts[0]
+        value = Number.parseFloat(mentionParts[1].slice(0, -1)) // chop off final ')' character
+        // logDebug('TMOcc:addOccurrence', `- found tagParts ${key} / ${value.toString()}`)
+      }
 
-    // if this has a numeric value add to total, taking into account that the day may have several values.
-    // $FlowFixMe[incompatible-type]
-    const prevValue: number = isNaN(this.valuesMap.get(dateStr)) ? 0 : this.valuesMap.get(dateStr)
-    if (!isNaN(value)) {
-      this.valuesMap.set(dateStr, prevValue + value)
-      this.count++
-      this.total += value
-      // logDebug('TMOccurrences:addOccurrence', `- ${key} / ${value} -> ${this.total} from ${this.count} on ${dateStr}`)
-    } else {
+      // if this has a numeric value add to total, taking into account that the day may have several values.
+      // $FlowFixMe[incompatible-type]
+      const prevValue: number = isNaN(this.valuesMap.get(dateStr)) ? 0 : this.valuesMap.get(dateStr)
+      if (!isNaN(value)) {
+        this.valuesMap.set(dateStr, prevValue + value)
+        this.count++
+        this.total += value
+        // logDebug('TMOcc:addOccurrence', `- ${key} / ${value} -> ${this.total} from ${this.count} on ${dateStr}`)
+      }
       // else just update the count
-      this.valuesMap.set(dateStr, prevValue + 1)
-      this.count++
-      this.total++
-      // logDebug('TMOccurrences:addOccurrence', `- ${key} increment -> ${this.total} from ${this.count} on ${dateStr}`)
+      else {
+        this.valuesMap.set(dateStr, prevValue + 1)
+        this.count++
+        this.total++
+        // logDebug('TMOcc:addOccurrence', `- ${key} increment -> ${this.total} from ${this.count} on ${dateStr}`)
+      }
+    }
+    catch (err) {
+      logError('TMOcc:addOccurrence', err.message)
     }
   }
 
@@ -183,7 +209,7 @@ export class TMOccurrences {
     for (let f of this.valuesMap.values()) {
       outArr.push(f)
     }
-    // logDebug('TMOccurrences:getValues', `for ${this.term} = ${outArr.toString()}`)
+    logDebug('TMOcc:getValues', `for ${this.term} = ${outArr.length} items: ${outArr.toString()}`)
     return outArr
   }
 
@@ -191,9 +217,9 @@ export class TMOccurrences {
    * Log all the details in the main valuesMap
    */
   logValuesMap(): void {
-    logDebug('TMOccurrences:logValuesMap', `- valuesMap for ${this.term}:`)
+    logDebug('TMOcc:logValuesMap', `- valuesMap for ${this.term}:`)
     this.valuesMap.forEach((v, k, m) => {
-      logDebug('TMOccurrences:logValuesMap', `  - ${k}: ${v}`)
+      logDebug('TMOcc:logValuesMap', `  - ${k}: ${v}`)
     })
   }
 
@@ -215,7 +241,7 @@ export class TMOccurrences {
         break
       }
       default: {
-        logError('TMOccurrences:getSparkline', `style '${style}' is not available`)
+        logError('TMOcc:getSparkline', `style '${style}' is not available`)
         break
       }
     }
@@ -229,47 +255,50 @@ export class TMOccurrences {
    */
   getStats(style: string): string {
     let output = ''
-    // logDebug('TMOccurrences:getStats', `starting for ${ this.term } type ${ this.type } style ${ style } `)
+    // logDebug('TMOcc:getStats', `starting for ${ this.term } type ${ this.type } style ${ style } `)
 
     const countStr = (!isNaN(this.count)) ? this.count.toLocaleString() : `none`
     const totalStr = (!isNaN(this.total) && this.total > 0) ? `total ${this.total.toLocaleString()}` : ''
     // This is the average per item, not the average per day. In general I feel this is more useful for numeric amounts
     const itemAvgStr = (!isNaN(this.total) && this.count > 0) ? `avg ${(this.total / this.count).toLocaleString([], { maximumSignificantDigits: 2 })}` : ''
-    // TODO: Decide how/when to use daily average instead of an item average
     // const dailyAvgStr = (!isNaN(this.total)) ? `dailyAvg ${(this.total / this.numDays).toLocaleString([], { maximumSignificantDigits: 2 })}` : ''
 
     switch (style) {
       case 'text': {
-        output = countStr // always start with count
-        if (this.count === this.total && this.type !== 'yesno') break // if it's a simple count, treat as such, even if that wasn't it's given type
-        switch (this.type) {
-          case 'yesno': {
-            output += " from " + this.numDays.toLocaleString()
-            break
-          }
-          case 'all': {
-            if (totalStr !== '') output += ", " + totalStr
-            if (itemAvgStr !== '') output += ", " + itemAvgStr
-            // if (dailyAvgStr !== '') output += ", " + dailyAvgStr
-            break
-          }
-          case 'total': {
-            output += ": " + totalStr
-            break
-          }
-          case 'average': {
-            if (itemAvgStr !== '') output += ": " + itemAvgStr
-            // if (dailyAvgStr !== '') output += ", " + dailyAvgStr
-            break
-          }
-          default: {
-            // nothing to add
+        // If we have no items, or simple single-unit counts, then just put count
+        if (this.total === this.count) {
+          output = `${countStr}`
+        }
+        else {
+          // Otherwise the output depends on the type
+          switch (this.type) {
+            case 'yesno': {
+              output = `${countStr} (from ${this.numDays})`
+              break
+            }
+            case 'total': {
+              output = `${totalStr} (from ${countStr})`
+              break
+            }
+            case 'average': {
+              if (itemAvgStr !== '') output += itemAvgStr
+              // if (dailyAvgStr !== '') output += ", " + dailyAvgStr
+              output += ` (from ${countStr})`
+              break
+            }
+            default: { // 'all'
+              if (totalStr !== '') output += totalStr
+              if (itemAvgStr !== '') output += ", " + itemAvgStr
+              // if (dailyAvgStr !== '') output += ", " + dailyAvgStr
+              output += ` (from ${countStr})`
+              break
+            }
           }
         }
         break
       }
       default: {
-        logError('TMOccurrences:getStats', `style '${style}' is not available`)
+        logError('TMOcc:getStats', `style '${style}' is not available`)
         output = '(unsupported style)'
         break
       }
@@ -280,7 +309,7 @@ export class TMOccurrences {
 
 /**
  * Gather all occurrences of requested hashtags and mentions for a given period.
- * Now also looks for requested 'progressYesNo' items too.
+ * Now also looks for requested 'progressYesNo', 'mentionTotal' and 'mentionAverage' items too.
  * Returns a list of TMOccurrences instances:
     term: string
     type: string // 'daily-average', 'item-average', 'total', 'yesno', 'count'
@@ -292,21 +321,23 @@ export class TMOccurrences {
  *
  * @author @jgclark
  * @param {string} periodString
- * @param {string} fromDateStr
- * @param {string} toDateStr
+ * @param {string} fromDateStr (YYYY-MM-DD)
+ * @param {string} toDateStr (YYYY-MM-DD)
  * @param {Array<string>} includedHashtags
  * @param {Array<string>} excludedHashtags
  * @param {Array<string>} includedMentions
  * @param {Array<string>} excludedMentions
  * @param {Array<string>} progressYesNo
+ * @param {Array<string>} includedMentionsTotal
+ * @param {Array<string>} includedMentionsAverage
  * @returns {Array<TMOccurrences>}
  */
-export function gatherOccurrences(periodString: string, fromDateStr: string, toDateStr: string, includedHashtags: Array<string>, excludedHashtags: Array<string>, includedMentions: Array<string>, excludedMentions: Array<string>, progressYesNo: Array<string> | string): Array<TMOccurrences> {
+export function gatherOccurrences(periodString: string, fromDateStr: string, toDateStr: string, includedHashtags: Array<string>, excludedHashtags: Array<string>, includedMentions: Array<string>, excludedMentions: Array<string>, progressYesNo: Array<string> | string, progressMentions: Array<string> | string, progressMentionsAverage: Array<string> | string, progressMentionsTotal: Array<string> | string): Array<TMOccurrences> {
   try {
 
-    logDebug('gatherOccurrences', `starting for ${periodString} (${fromDateStr}-${toDateStr})`)
+    logDebug('gatherOccurrences', `starting for ${periodString} (${fromDateStr} - ${toDateStr})`)
     const periodDailyNotes = DataStore.calendarNotes.filter(
-      (p) => withinDateRange(getDateStringFromCalendarFilename(p.filename), fromDateStr, toDateStr))
+      (p) => withinDateRange(getDateStringFromCalendarFilename(p.filename), unhyphenateString(fromDateStr), unhyphenateString(toDateStr)))
     if (periodDailyNotes.length === 0) {
       logWarn('gatherOccurrences', `no matching daily notes found between ${fromDateStr} and ${toDateStr}`)
       return [] // for completeness
@@ -320,7 +351,7 @@ export function gatherOccurrences(periodString: string, fromDateStr: string, toD
 
     //------------------------------
     // Review each wanted YesNo type
-    const YesNoListArr = (typeof progressYesNo === 'string') ? progressYesNo.split(',') : progressYesNo
+    const YesNoListArr = (typeof progressYesNo === 'string') ? progressYesNo.split(',') : progressYesNo // make sure this is an array first
     for (let wantedItem of YesNoListArr) {
       // initialise a new TMOccurence for this YesNo item
       const thisOcc = new TMOccurrences(wantedItem, 'yesno', fromDateStr, toDateStr)
@@ -402,10 +433,22 @@ export function gatherOccurrences(periodString: string, fromDateStr: string, toD
 
     //------------------------------
     // Now repeat for @mentions
-    // TODO: Add exclusions -- as section above.
-    for (let wantedMention of includedMentions) {
+    // TODO: Add exclusions -- as section above
+    // There are now 3 kinds of @mentions to process: make a superset of them to sort and then process in one go
+    // Make sure they are arrays first.
+    const allMentionsArr = stringListOrArrayToArray(progressMentions, ',')
+    const averageMentionsArr = stringListOrArrayToArray(progressMentionsAverage, ',')
+    const totalMentionsArr = stringListOrArrayToArray(progressMentionsTotal, ',')
+    const combinedMentions = []
+    allMentionsArr.forEach((m) => { combinedMentions.push([m, 'all']) })
+    averageMentionsArr.forEach((m) => { combinedMentions.push([m, 'average']) })
+    totalMentionsArr.forEach((m) => { combinedMentions.push([m, 'total']) })
+    combinedMentions.sort()
+
+    for (let thisMention of combinedMentions) {
       // initialise a new TMOccurence for this mention
-      const thisOcc = new TMOccurrences(wantedMention, 'all', fromDateStr, toDateStr)
+      const [thisName, thisType] = thisMention
+      const thisOcc = new TMOccurrences(thisName, thisType, fromDateStr, toDateStr)
 
       // For each daily note in the period, look at each mention in reverse order to make subset checking work
       for (const n of periodDailyNotes) {
@@ -413,8 +456,9 @@ export function gatherOccurrences(periodString: string, fromDateStr: string, toD
         const seenMentions = n.mentions.slice().reverse()
         let lastMention = ''
         for (const mention of seenMentions) {
+
           const mentionWithoutNumberPart = (mention.split('(', 1))[0]
-          // logDebug('gatherOccurrences', `- reviewing ${mention} [${mentionWithoutNumberPart}] looking for ${wantedMention} on ${thisDateStr}`)
+          // logDebug('gatherOccurrences', `- reviewing ${mention} [${mentionWithoutNumberPart}] looking for ${thisName} on ${thisDateStr}`)
           // if this tag is starting subset of the last one, assume this is an example of the issue, so skip this tag
           if (caseInsensitiveStartsWith(mentionWithoutNumberPart, lastMention)) {
             // logDebug('gatherOccurrences', `- Found ${mention} but ignoring as part of a longer mention of the same name`)
@@ -422,14 +466,14 @@ export function gatherOccurrences(periodString: string, fromDateStr: string, toD
           }
           else {
             // check this is on inclusion, or not on exclusion list, before adding
-            if (caseInsensitiveMatch(mentionWithoutNumberPart, wantedMention)) {
+            if (caseInsensitiveMatch(mentionWithoutNumberPart, thisName)) {
               // logDebug('gatherOccurrences', `- Found matching occurrence ${mention} on date ${n.filename}`)
               thisOcc.addOccurrence(mention, thisDateStr)
             } else {
               // logDebug('gatherOccurrences', `- x ${mention} not wanted`)
             }
           }
-          lastMention = mention
+          lastMention = thisName
         }
       }
       tmOccurrencesArr.push(thisOcc)
@@ -453,11 +497,12 @@ export function gatherOccurrences(periodString: string, fromDateStr: string, toD
  * @param {string} toDateStr 
  * @param {string} style 
  * @param {boolean} showSparklines 
+ * @param {boolean} sortOutput 
  * @returns Array<string>
  */
-export function generateProgressUpdate(occObjs: Array<TMOccurrences>, periodString: string, fromDateStr: string, toDateStr: string, style: string, showSparklines: boolean): Array<string> {
+export function generateProgressUpdate(occObjs: Array<TMOccurrences>, periodString: string, fromDateStr: string, toDateStr: string, style: string, showSparklines: boolean, sortOutput: boolean): Array<string> {
   try {
-    logDebug('generateProgressUpdate', `starting for ${periodString} (${fromDateStr}-${toDateStr}) with ${occObjs.length} occObjs`)
+    logDebug('generateProgressUpdate', `starting for ${periodString} (${fromDateStr} - ${toDateStr}) with ${occObjs.length} occObjs`)
 
     // Get length of longest progress term (to use with sparklines)
     const maxTermLen = Math.max(...occObjs.map((m) => m.term.length))
@@ -483,6 +528,10 @@ export function generateProgressUpdate(occObjs: Array<TMOccurrences>, periodStri
         }
       }
       outputArray.push(thisOutput)
+      if (sortOutput) {
+        // sort using locale-aware sorting
+        outputArray.sort((a, b) => a.localeCompare(b))
+      }
     }
     return outputArray
   }